scripts: Fix deleteOldComments.sh
 * Get comments from github-actions, not only our android bot
 * Fix count (was broken for 0 comments) and only try to delete if count > 0

+++ scripts/deleteOldComments.sh
@@ -1,11 +1,8 @@
 #!/usr/bin/env bash
-#1: LOG_USERNAME
-#2: LOG_PASSWORD
-#3: DRONE_BUILD_NUMBER
-#4: BRANCH (stable or master)
-#5: TYPE (IT or Unit)
-#6: DRONE_PULL_REQUEST
-#7: GITHUB_TOKEN
+#1: BRANCH
+#2: TYPE
+#3: PR
+#4: GITHUB_TOKEN
 
 BRANCH=$1
 TYPE=$2
@@ -16,13 +13,15 @@ BRANCH_TYPE=$BRANCH-$TYPE
 
  # delete all old comments, matching this type
 echo "Deleting old comments for $BRANCH_TYPE"
-oldComments=$(curl 2>/dev/null --header "authorization: Bearer $GITHUB_TOKEN" -X GET https://api.github.com/repos/nextcloud/android/issues/$PR/comments | jq --arg TYPE $BRANCH_TYPE '.[] | (.id |tostring) + "|" + (.user.login | test("nextcloud-android-bot") | tostring) + "|" + (.body | test([$TYPE]) | tostring)'| grep "true|true" | tr -d "\"" | cut -f1 -d"|")
-count=$(echo $oldComments | grep true | wc -l)
+oldComments=$(curl 2>/dev/null --header "authorization: Bearer $GITHUB_TOKEN" -X GET https://api.github.com/repos/nextcloud/android/issues/$PR/comments | jq --arg TYPE $BRANCH_TYPE '.[] | (.id |tostring) + "|" + (.user.login | test("(nextcloud-android-bot|github-actions)") | tostring) + "|" + (.body | test([$TYPE]) | tostring)'| grep "true|true" | tr -d "\"" | cut -f1 -d"|")
+count=$(echo -n "$oldComments" | grep -c '^')
 echo "Found $count old comments"
 
-echo $oldComments | while read comment ; do
-echo "Deleting comment: $comment"
-curl 2>/dev/null --header "authorization: Bearer $GITHUB_TOKEN" -X DELETE https://api.github.com/repos/nextcloud/android/issues/comments/$comment
-done
+if [ "$count" -gt 0 ]; then
+  echo "$oldComments" | while read comment ; do
+    echo "Deleting comment: $comment"
+    curl 2>/dev/null --header "authorization: Bearer $GITHUB_TOKEN" -X DELETE https://api.github.com/repos/nextcloud/android/issues/comments/$comment
+  done
+fi
 
 exit 0
